refacto et ajout de fichier service + impl
