+++ src/main/java/com/materio/materio_backend/jpa/entity/user/EnumRole.java
@@ -2,6 +2,5 @@ package com.materio.materio_backend.jpa.entity.user;
 
 public enum EnumRole {
     ROLE_USER,
-    ROLE_ADMIN,
-    ROLE_MODERATOR
+    ROLE_ADMIN
 }
